test-requirements.txt
azure-functions
flask
fastapi
azurefunctions-extensions-http-fastapi

pytest
pytest-coverage
ddt
pyyaml
gunicorn
pyfir>=1.0.0